Game settings initial work
