+++ apps/enshrouded/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2024"
 repository = "https://github.com/mbround18/enshrouded-docker"
 
 [dependencies]
-clap = { version = "4.5.35", features = ["derive"] }
+clap = { version = "4.5.41", features = ["derive"] }
 gsm-instance = {path = "../../libs/gsm-instance"}
 gsm-cron = {path = "../../libs/gsm-cron"}
 gsm-shared = {path = "../../libs/gsm-shared"}
@@ -13,9 +13,9 @@ gsm-monitor = {path = "../../libs/gsm-monitor"}
 gsm-notifications = {path ="../../libs/gsm-notifications"}
 env-parse = {path = "../../libs/env-parse"}
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
+serde_json = "1.0.141"
 tracing = "0.1"
 tracing-subscriber = "0.3"
-tokio = { version = "1.44.1", features = ["rt", "rt-multi-thread", "macros"] }
+tokio = { version = "1.46.1", features = ["rt", "rt-multi-thread", "macros"] }
 regex = "1.11.1"
 lazy_static = "1.5.0"
