+++ apps/enshrouded/src/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod config_io;
+pub mod env_overrides;
 mod extract_player_name;
 
 pub use extract_player_name::*;
