+++ libs/gsm-backup/Cargo.toml
@@ -11,4 +11,4 @@ tar = "0.4"
 thiserror = "2"
 
 [dev-dependencies]
-tempfile = "3.19"
+tempfile = "3.20"
