+++ libs/gsm-cron/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2024"
 [dependencies]
 tracing = "0"
 cron = "0"
-chrono = "0.4.40"
-tokio = "1.44.1"
+chrono = "0.4.41"
+tokio = "1.46.1"
 gsm-shared = {path = "../gsm-shared"}
-nix = { version = "0.29.0", features = ["signal"] }
+nix = { version = "0.30.1", features = ["signal"] }
