+++ libs/gsm-instance/Cargo.toml
@@ -8,10 +8,10 @@ daemonize = "0"
 serde = { version = "1.0.219", features = ["derive"] }
 thiserror = "2.0.12"
 tracing = "0.1"
-tempfile = "3.19.1"
-which = "7.0.2"
+tempfile = "3.20.0"
+which = "8.0.0"
 sysinfo = "0"
 regex = "1.11.1"
 strsim = "0"
-tokio = { version="1.44.1", features = ["full", "process"] }
-nix = "0.29.0"
+tokio = { version="1.46.1", features = ["full", "process"] }
+nix = "0.30.1"
