+++ libs/gsm-mod-manager/Cargo.toml
@@ -9,9 +9,9 @@ path = "src/lib.rs"
 
 [dependencies]
 tracing = "0.1"
-tempfile = "3.19.1"
+tempfile = "3.20.0"
 walkdir = "2.5.0"
-zip = "2.5.0"
+zip = "4.3.0"
 fs_extra = "1"
 thiserror = "2"
 regex = "1"
