+++ libs/gsm-notifications/Cargo.toml
@@ -4,8 +4,8 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-reqwest = "0.12.15"
+reqwest = "0.12.22"
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
+serde_json = "1.0.141"
 gsm-shared = {path = "../gsm-shared"}
 tracing = "0"
