+++ libs/gsm-serde/Cargo.toml
@@ -5,5 +5,5 @@ edition = "2024"
 
 [dependencies]
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
+serde_json = "1.0.141"
 ini-derive = {path = "../ini-derive"}
