+++ libs/gsm-shared/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2024"
 [dependencies]
 serde = {version = "1",features = ["derive", "default"] }
 tracing = "0.1"
-tempfile = "3.19.1"
+tempfile = "3.20.0"
 walkdir = "2.5.0"
 reqwest = {version = "0", features = ["json", "default-tls", "blocking"]}
 cached = "0"
-md5 = "0.7"
+md5 = "0.8"
