+++ tools/env-parser/Cargo.toml
@@ -8,9 +8,9 @@ name = "env-parser"
 path = "src/main.rs"
 
 [dependencies]
-clap = { version = "4.5.35", features = ["derive"] }
+clap = { version = "4.5.41", features = ["derive"] }
 tera = "1.20.0"
 regex = "1"
 serde = { version = "1.0.219", features = ["derive"] }
-serde_json = "1.0.140"
+serde_json = "1.0.141"
